% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dshSophiaCreateBaseline.R
\name{dshSophiaCreateBaseline}
\alias{dshSophiaCreateBaseline}
\title{Create a baseline data frame on the federated node}
\usage{
dshSophiaCreateBaseline(procedure_id = NULL)
}
\value{
A federated data frame named 'baseline'.
}
\description{
Given a vector of valid measurement table Concept IDs the function creates a 'baseline' data frame on the federated node. Here, 'baseline' means the first measurement available, and as such may be different for different Concept IDs. The function also creates an approximate 'age_baseline' variable and merged the baseline data with gender if available in the person table.
}
\examples{
\dontrun{
# connect to the federated system
dshSophiaConnect()

# load database resources
dshSophiaLoad()

# create a 'baseline' data frame on the federated node
dshSophiaCreateBaseline(concept_id = c(3038553, 3025315, 37020574))

# check result
dsBaseClient::ds.summary("baseline")
}
}
