% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dshSophiaGetCor.R
\name{dshSophiaGetCor}
\alias{dshSophiaGetCor}
\title{Get the correlation between two continuous variables}
\usage{
dshSophiaGetCor(
  x,
  y,
  method = "pearson",
  keep_procedure = NA,
  remove_procedure = NA,
  keep_observation = NA,
  remove_observation = NA,
  reverse_x = FALSE,
  reverse_y = FALSE
)
}
\arguments{
\item{x}{A character, must correspond to the name of a numerical column in 'baseline'.}

\item{y}{A character, must correspond to the name of a numerical column in 'baseline'.}

\item{method}{A character, specifying the type of correlation. Currently only 'pearson' is allowed and is thus default.}

\item{keep_procedure}{A numeric, must be a valid Concept ID for a procedure column created using 'dshSophiaCreateBaseline'. Will only keep rows with this procedure. Defaults to NA.}

\item{remove_procedure}{A numeric, must be a valid Concept ID for a procedure column created using 'dshSophiaCreateBaseline'. Will remove all rows with this procedure. Defaults to NA.}

\item{keep_observation}{A numeric, must be a valid Concept ID for an observation column created using 'dshSophiaCreateBaseline'. Will only keep rows with this observation. Defaults to NA.}

\item{remove_observation}{A numeric, must be a valid Concept ID for an observation column created using 'dshSophiaCreateBaseline'. Will remove all rows with this observation. Defaults to NA.}

\item{reverse_x}{A Boolean, set to TRUE if x should be reversed. Defaults to FALSE.}

\item{reverse_y}{A Boolean, set to TRUE if y should be reversed. Defaults to FALSE.}
}
\value{
A data frame with the estimated correlation and associated metrics. If privacy checks are not passed, a data frame with NAs will be returned.
}
\description{
Given the names of two numeric columns in 'baseline', calculates and returns the Pearson correlation and associated metrics. Several checks are performed to make sure the function does not violate privacy rules. The user has the option to subset data by procedure and observation prior to estimating the correlation. The variables may also be reversed, which is useful when negative values are considered 'good', such as percent weight loss.
}
\examples{
\dontrun{
# connect to the federated system
dshSophiaConnect(include = "abos")

# load database resources
dshSophiaLoad()

# create a 'baseline' data frame on the federated node, 
calculate age at first visit and include 'bypass of stomach' procedure and 'type 2 diabetes' observation
dshSophiaCreateBaseline(age_at_first = "visit", procedure_id = 40483096, observation_id = 201826)
   
# add longitudinal weight measures
dshSophiaMergeLongMeas(concept_id = 3038553)

# get correlation between age and weight for subgroup with bypass and T2D
dshSophiaGetCor(x = "t1_3038553", y = "age_at_first_visit", keep_procedure = 40483096, keep_observation = 201826)
}
}
