% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dshSophiaLoad.R
\name{dshSophiaLoad}
\alias{dshSophiaLoad}
\title{Load database resources}
\usage{
dshSophiaLoad()
}
\value{
Nothing, all databases are loaded locally.
}
\description{
Loads all database resources on federated nodes.
}
\examples{
\dontrun{
# first we must connect to the federated system
dshSophiaConnect()

# now we can load database resources
dshSophiaLoad()
}
}
