% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dshSophiaOverview.R
\name{dshSophiaOverview}
\alias{dshSophiaOverview}
\title{Get an overview of available data}
\usage{
dshSophiaOverview()
}
\value{
A data frame.
}
\description{
Runs a query that iterates over every cohort and gathers information about the variables available in each of the cohort's OMOP tables.
}
\examples{
\dontrun{
# connect to the federated system
dshSophiaConnect()

# load database resources
dshSophiaLoad()

# generate data overview
overview <- dshSophiaOverview()
}
}
