% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dshSophiaConnect.R
\name{dshSophiaConnect}
\alias{dshSophiaConnect}
\title{Connect to SOPHIA nodes}
\usage{
dshSophiaConnect(
  username = Sys.getenv("fdb_user"),
  password = Sys.getenv("fdb_password"),
  include = NULL,
  exclude = NULL
)
}
\arguments{
\item{username}{A character; your username. Defaults to \code{Sys.getenv("fdb_user")}.}

\item{password}{A character; your password. Defaults to \code{Sys.getenv("fdb_password")}.}

\item{include}{A character or vector of characters containing only the nodes you want to connect to.}

\item{exclude}{A character or vector of characters containing only the nodes you do not want to connect to.}
}
\value{
An Opals object (\code{opals}) is assigned to the Global environment.
}
\description{
Allows the user to connect to the federated SOPHIA nodes, with the option of including and excluding specific nodes. Note that there are two rounds of connecting: First the user connects to each node in order to retrieve a list of all cohorts included at each node. Then, the user connects to each cohort separately.
}
\examples{
\dontrun{
# connect to all nodes, assuming username and password is specified in .Renvironment:
dshSophiaConnect()

# connect to all nodes, manually providing username and password:
dshSophiaConnect(username = "user1", password = "pass1")

# include specific nodes:
dshSophiaConnect(include = c("node1", "node2"))

# exclude specific nodes:
dshSophiaConnect(exclude = c("node3", "node4"))
}
}
