% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dshSophiaConnect.R
\name{dshSophiaConnect}
\alias{dshSophiaConnect}
\title{Connect to SOPHIA nodes}
\usage{
dshSophiaConnect(
  username = Sys.getenv("fdb_user"),
  password = Sys.getenv("fdb_password"),
  include = NULL,
  exclude = NULL
)
}
\arguments{
\item{username}{A character; your username. Defaults to \code{Sys.getenv("fdb_user")}.}

\item{password}{A character; your password. Defaults to \code{Sys.getenv("fdb_password")}.}

\item{include}{A character or vector of characters containing only the nodes you want to connect to.}

\item{exclude}{A character or vector of characters containing only the nodes you do not want to connect to.}
}
\value{
An Opals object (\code{opals}) is assigned to the Global environment.
}
\description{
Allows the user to connect to the SOPHIA federated database, with the option of including and excluding specific nodes (a node in this context is a server that hosts a database). Note that connecting to the federated database is a two-step process: First, the user connects to each individual node in order to retrieve a list of all cohorts that are hosted on that specific node (cohorts in this context refer to a specific dataset associated with a study or research project). Then, the user is disconnected, and then reconnects to each individual cohort.
}
\examples{
\dontrun{
# connect to all nodes, assuming username and password is specified in .Renviron:
dshSophiaConnect()

# connect to all nodes, manually providing username and password:
dshSophiaConnect(username = "username", password = "password")

# include specific nodes:
dshSophiaConnect(include = c("node1", "node2"))

# exclude specific nodes:
dshSophiaConnect(exclude = c("node3", "node4"))
}
}
