% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dshSophiaLoad.R
\name{dshSophiaLoad}
\alias{dshSophiaLoad}
\title{Load SOPHIA federated database resources}
\usage{
dshSophiaLoad()
}
\value{
Nothing, all databases are loaded locally on each federated node.
}
\description{
Loads all database resources on federated nodes.
}
\examples{
\dontrun{
# connect to the federated system
dshSophiaConnect(include = "abos")

# load database resources
dshSophiaLoad()
}
}
