% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dshSophiaMeasureDesc.R
\name{dshSophiaMeasureDesc}
\alias{dshSophiaMeasureDesc}
\title{Get descriptive information about a variable in the measurement table}
\usage{
dshSophiaMeasureDesc(
  variable,
  keep_procedure = NA,
  remove_procedure = NA,
  keep_observation = NA,
  remove_observation = NA,
  keep_gender = NA,
  remove_gender = NA
)
}
\value{
A data frame.
}
\description{
Given a valid Concept ID the function gathers descriptive information about the corresponding variable in the measurement table and outputs a summary of the results. If the data is longitudinal the output will contain one row per time point.
}
\examples{
\dontrun{
# connect to the federated system
dshSophiaConnect()

# load database resources
dshSophiaLoad()

# get descriptive information about BMI
df <- dshSophiaMeasureDesc(concept_id = 3038553)
}
}
