% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dshSophiaShow.R
\name{dshSophiaShow}
\alias{dshSophiaShow}
\title{Show available nodes and cohorts}
\usage{
dshSophiaShow(
  username = Sys.getenv("fdb_user"),
  password = Sys.getenv("fdb_password")
)
}
\arguments{
\item{username}{A character, your username. Defaults to \code{Sys.getenv("fdb_user")}.}

\item{password}{A character, your password. Defaults to \code{Sys.getenv("fdb_password")}.}
}
\value{
A dataframe with all available nodes and cohorts in 'long' format, along with date if creation and last update.
}
\description{
Downloads an up-to-date .csv file from SIB containing node status and URL, then
via a two-step process logs in to each node to retrive cohort names. The output
is a dataframe containing all available nodes and cohorts.
}
\examples{
\dontrun{
# show all nodes and cohorts, assuming username and password is specified in .Renvironment:
dshSophiaShow()

# show all nodes and cohorts, manually providing username and password:
dshSophiaShow(username = "user1", password = "pass1")
}
}
